feat: new withSyncToWebStorage plus docs and small fixes
- Added new withSyncToWebStorage
- Docs for all functions
- Fixed clear of selection after filter changes or sort changed when paginated, for multi and single section
