+++ libs/ngrx-traits/signals/src/lib/index.ts
@@ -1,12 +1,18 @@
-export * from './with-entities-sort/with-entities-sort.utils';
 export * from './util';
 export * from './with-call-status/with-call-status';
-export * from './with-entities-filter/with-entities-filter';
-export * from './with-entities-local-pagination/with-entities-local-pagination';
-export * from './with-entities-remote-pagination/with-entities-remote-pagination';
-export * from './with-entities-sort/with-entities-sort';
+export * from './with-entities-filter/with-entities-local-filter';
+export * from './with-entities-filter/with-entities-remote-filter';
+export * from './with-entities-pagination/with-entities-local-pagination';
+export * from './with-entities-pagination/with-entities-remote-pagination';
+export {
+  Sort,
+  SortDirection,
+} from './with-entities-sort/with-entities-sort.utils';
+export * from './with-entities-sort/with-entities-local-sort';
+export * from './with-entities-sort/with-entities-remote-sort';
 export * from './with-entities-single-selection/with-entities-single-selection';
 export * from './with-entities-multi-selection/with-entities-multi-selection';
 export * from './with-entities-loading-call/with-entities-loading-call';
 export * from './with-logger/with-logger';
 export * from './with-calls/with-calls';
+export * from './with-sync-to-web-storage/with-sync-to-web-storage';
