+++ libs/ngrx-traits/signals/src/lib/with-call-status/with-call-status.ts
@@ -47,6 +47,34 @@ export type NamedCallStateMethods<Prop extends string> = {
   [K in Prop as `set${Capitalize<string & K>}Error`]: () => void;
 };
 
+/**
+ * Generates necessary state, computed and methods for call progress status to the store
+ * @param config - Configuration object
+ * @param config.prop - The name of the property for which this represents the call status
+ * @param config.initialValue - The initial value of the call status
+ * @param config.collection - The name of the collection for which this represents the call status is an alias to prop param
+ * they do the same thing
+ *
+ * prop or collection is required
+ * @example
+ * const store = signalStore(
+ *  withCallStatus({ prop: 'users', })
+ *  // other valid configurations
+ *  // withCallStatus()
+ *  // withCallStatus({ collection: 'users', initialValue: 'loading' })
+ *  )
+ *
+ *  // generates the following signals
+ *  store.usersCallStatus // 'init' | 'loading' | 'loaded' | { error: unknown }
+ *  // generates the following computed signals
+ *  store.usersLoading // boolean
+ *  store.usersLoaded // boolean
+ *  store.usersError // unknown | null
+ *  // generates the following methods
+ *  store.setUsersLoading // () => void
+ *  store.setUsersLoaded // () => void
+ *  store.setUsersError // (error?: unknown) => void
+ */
 export function withCallStatus(config?: {
   initialValue?: CallStatus;
 }): SignalStoreFeature<
@@ -57,10 +85,46 @@ export function withCallStatus(config?: {
     methods: CallStateMethods;
   }
 >;
-export function withCallStatus<Prop extends string>(config?: {
-  prop: Prop;
-  initialValue?: CallStatus;
-}): SignalStoreFeature<
+
+/**
+ * Generates necessary state, computed and methods for call progress status to the store
+ * @param config - Configuration object
+ * @param config.prop - The name of the property for which this represents the call status
+ * @param config.initialValue - The initial value of the call status
+ * @param config.collection - The name of the collection for which this represents the call status is an alias to prop param
+ * they do the same thing
+ *
+ * prop or collection is required
+ * @example
+ * const store = signalStore(
+ *  withCallStatus({ prop: 'users', })
+ *  // other valid configurations
+ *  // withCallStatus()
+ *  // withCallStatus({ collection: 'users', initialValue: 'loading' })
+ *  )
+ *
+ *  // generates the following signals
+ *  store.usersCallStatus // 'init' | 'loading' | 'loaded' | { error: unknown }
+ *  // generates the following computed signals
+ *  store.usersLoading // boolean
+ *  store.usersLoaded // boolean
+ *  store.usersError // unknown | null
+ *  // generates the following methods
+ *  store.setUsersLoading // () => void
+ *  store.setUsersLoaded // () => void
+ *  store.setUsersError // (error?: unknown) => void
+ */
+export function withCallStatus<Prop extends string>(
+  config?:
+    | {
+        prop: Prop;
+        initialValue?: CallStatus;
+      }
+    | {
+        collection: Prop;
+        initialValue?: CallStatus;
+      },
+): SignalStoreFeature<
   { state: {}; signals: {}; methods: {} },
   {
     state: NamedCallState<Prop>;
@@ -71,10 +135,13 @@ export function withCallStatus<Prop extends string>(config?: {
 export function withCallStatus<Prop extends string>({
   prop,
   initialValue = 'init',
+  collection,
 }: {
   prop?: Prop;
+  collection?: Prop;
   initialValue?: CallStatus;
 } = {}): SignalStoreFeature {
+  prop = prop ?? collection;
   const {
     callStatusKey,
     errorKey,
