+++ libs/ngrx-traits/signals/src/lib/with-entities-filter/with-entities-filter.util.ts
@@ -1,3 +1,14 @@
+import { Signal } from '@angular/core';
+import {
+  concatMap,
+  debounce,
+  distinctUntilChanged,
+  EMPTY,
+  of,
+  pipe,
+  timer,
+} from 'rxjs';
+
 import { capitalize } from '../util';
 
 export function getWithEntitiesFilterKeys(config?: { collection?: string }) {
@@ -9,4 +20,30 @@ export function getWithEntitiesFilterKeys(config?: { collection?: string }) {
       ? `filter${capitalizedProp}Entities`
       : 'filterEntities',
   };
+}
+
+export function debounceFilterPipe<Filter, Entity>(filter: Signal<Filter>) {
+  return pipe(
+    debounce(
+      (value: {
+        filter: Filter;
+        debounce?: number;
+        patch?: boolean;
+        forceLoad?: boolean;
+      }) => (value?.forceLoad ? EMPTY : timer(value.debounce || 300)),
+    ),
+    concatMap((payload) =>
+      payload.patch
+        ? of({
+            ...payload,
+            filter: { ...filter?.(), ...payload?.filter },
+          })
+        : of(payload),
+    ),
+    distinctUntilChanged(
+      (previous, current) =>
+        !current?.forceLoad &&
+        JSON.stringify(previous?.filter) === JSON.stringify(current?.filter),
+    ),
+  );
 }
