+++ libs/ngrx-traits/signals/src/lib/with-entities-single-selection/with-entities-single-selection.ts
@@ -1,9 +1,10 @@
-import { computed, Signal } from '@angular/core';
+import { computed, effect, Signal } from '@angular/core';
 import {
   patchState,
   signalStoreFeature,
   SignalStoreFeature,
   withComputed,
+  withHooks,
   withMethods,
   withState,
 } from '@ngrx/signals';
@@ -16,6 +17,9 @@ import {
 import type { StateSignal } from '@ngrx/signals/src/state-signal';
 
 import { capitalize, getWithEntitiesKeys } from '../util';
+import { getWithEntitiesFilterKeys } from '../with-entities-filter/with-entities-filter.util';
+import { getWithEntitiesLocalPaginationKeys } from '../with-entities-pagination/with-entities-local-pagination.util';
+import { getWithEntitiesSortKeys } from '../with-entities-sort/with-entities-sort.util';
 
 export type EntitiesSingleSelectionState = {
   selectedId?: string | number;
@@ -72,6 +76,37 @@ function getEntitiesSingleSelectionKeys(config?: { collection?: string }) {
   };
 }
 
+/**
+ * Generates state, computed and methods for single selection of entities. Requires withEntities to be present before this function.
+ * @param config
+ * @param config.collection - The collection name
+ * @param config.entity - The entity type
+ *
+ * @example
+ * const entity = type<Product>();
+ * const collection = 'products';
+ * export const store = signalStore(
+ *   { providedIn: 'root' },
+ *   // Required withEntities and withCallStatus
+ *   withEntities({ entity, collection }),
+ *   withCallStatus({ prop: collection, initialValue: 'loading' }),
+ *
+ *   withEntitiesSingleSelection({
+ *     entity,
+ *     collection,
+ *   }),
+ *  );
+ *
+ *  // generates the following signals
+ *  store.productsSelectedId // string | number | undefined
+ *  // generates the following computed signals
+ *  store.productsSelectedEntity // Entity | undefined
+ *  // generates the following methods
+ *  store.selectProductEntity // (options: { id: string | number }) => void
+ *  store.deselectProductEntity // (options: { id: string | number }) => void
+ *  store.toggleProductEntity // (options: { id: string | number }) => void
+ */
+
 export function withEntitiesSingleSelection<
   Entity extends { id: string | number },
 >(options: {
@@ -88,6 +123,36 @@ export function withEntitiesSingleSelection<
     methods: EntitiesSingleSelectionMethods;
   }
 >;
+/**
+ * Generates state, computed and methods for single selection of entities. Requires withEntities to be present before this function.
+ * @param config
+ * @param config.collection - The collection name
+ * @param config.entity - The entity type
+ *
+ * @example
+ * const entity = type<Product>();
+ * const collection = 'products';
+ * export const store = signalStore(
+ *   { providedIn: 'root' },
+ *   // Required withEntities and withCallStatus
+ *   withEntities({ entity, collection }),
+ *   withCallStatus({ prop: collection, initialValue: 'loading' }),
+ *
+ *   withEntitiesSingleSelection({
+ *     entity,
+ *     collection,
+ *   }),
+ *  );
+ *
+ *  // generates the following signals
+ *  store.productsSelectedId // string | number | undefined
+ *  // generates the following computed signals
+ *  store.productsSelectedEntity // Entity | undefined
+ *  // generates the following methods
+ *  store.selectProductEntity // (options: { id: string | number }) => void
+ *  store.deselectProductEntity // (options: { id: string | number }) => void
+ *  store.toggleProductEntity // (options: { id: string | number }) => void
+ */
 export function withEntitiesSingleSelection<
   Entity extends { id: string | number },
   Collection extends string,
@@ -116,6 +181,9 @@ export function withEntitiesSingleSelection<
   collection?: Collection;
 }): SignalStoreFeature<any, any> {
   const { entityMapKey } = getWithEntitiesKeys(config);
+  const { filterKey } = getWithEntitiesFilterKeys(config);
+  const { sortKey } = getWithEntitiesSortKeys(config);
+  const { paginationKey } = getWithEntitiesLocalPaginationKeys(config);
   const {
     selectedEntityKey,
     selectEntityKey,
@@ -159,5 +227,39 @@ export function withEntitiesSingleSelection<
         },
       };
     }),
+    withHooks((store) => ({
+      onInit: () => {
+        // we need reset the selections if to 0 when the filter changes or sorting changes and is paginated
+        if (filterKey in store || sortKey in store) {
+          const filter = store[filterKey] as Signal<unknown>;
+          const sort = store[sortKey] as Signal<unknown>;
+          const deselectEntity = store[deselectEntityKey] as () => void;
+          let lastFilter = filter?.();
+          let lastSort = sort?.();
+          /** TODO: there is a small problem here when used together with withSyncToWebStorage an filter or sorting
+           the stored selection will get cleared because this logic detects the sorting and the filtering changing
+           from the default value to the stored value and resets the selection */
+          effect(
+            () => {
+              if (
+                (store[paginationKey] && lastSort !== sort?.()) ||
+                lastFilter !== filter?.()
+              ) {
+                console.log(
+                  'resetting selection',
+                  store[paginationKey] && lastSort !== sort?.(),
+                  lastFilter !== filter?.(),
+                  { filter: filter?.(), sort: sort?.(), lastFilter, lastSort },
+                );
+                lastFilter = filter?.();
+                lastSort = sort?.();
+                deselectEntity();
+              }
+            },
+            { allowSignalWrites: true },
+          );
+        }
+      },
+    })),
   );
 }
