+++ libs/ngrx-traits/signals/src/lib/with-logger/with-logger.ts
@@ -1,6 +1,10 @@
 import { effect } from '@angular/core';
 import { getState, signalStoreFeature, withHooks } from '@ngrx/signals';
 
+/**
+ * Log the state of the store on every change
+ * @param name - The name of the store to log
+ */
 export function withLogger(name: string) {
   return signalStoreFeature(
     withHooks({
