+++ package.json
@@ -33,13 +33,13 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.0",
-    "@angular/cdk": "17.3.0",
-    "@angular/common": "17.3.0",
-    "@angular/core": "17.3.0",
-    "@angular/forms": "17.3.0",
-    "@angular/material": "^17.3.0",
-    "@angular/platform-browser": "17.3.0",
+    "@angular/animations": "17.3.3",
+    "@angular/cdk": "17.3.3",
+    "@angular/common": "17.3.3",
+    "@angular/core": "17.3.3",
+    "@angular/forms": "17.3.3",
+    "@angular/material": "^17.3.3",
+    "@angular/platform-browser": "17.3.3",
     "@ngrx/effects": "17.1.1",
     "@ngrx/entity": "17.1.1",
     "@ngrx/signals": "^17.1.1",
@@ -52,21 +52,22 @@
     "zone.js": "0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.0",
-    "@angular-devkit/core": "17.3.0",
-    "@angular-devkit/schematics": "17.3.0",
+    "@angular-devkit/build-angular": "17.3.3",
+    "@angular-devkit/core": "17.3.3",
+    "@angular-devkit/schematics": "17.3.3",
     "@angular-eslint/eslint-plugin": "17.3.0",
     "@angular-eslint/eslint-plugin-template": "17.3.0",
     "@angular-eslint/template-parser": "17.3.0",
-    "@angular/cli": "~17.3.0",
-    "@angular/compiler": "17.3.0",
-    "@angular/compiler-cli": "17.3.0",
-    "@angular/language-service": "17.3.0",
-    "@angular/platform-browser-dynamic": "17.3.0",
-    "@angular/router": "17.3.0",
+    "@angular/cli": "~17.3.3",
+    "@angular/compiler": "17.3.3",
+    "@angular/compiler-cli": "17.3.3",
+    "@angular/language-service": "17.3.3",
+    "@angular/platform-browser-dynamic": "17.3.3",
+    "@angular/router": "17.3.3",
     "@babel/cli": "^7.17.10",
     "@babel/core": "^7.18.0",
     "@babel/plugin-proposal-class-properties": "^7.17.12",
+    "@babel/plugin-proposal-object-rest-spread": "^7.17.12",
     "@babel/plugin-proposal-decorators": "^7.17.12",
     "@babel/preset-env": "^7.18.0",
     "@babel/preset-typescript": "^7.17.12",
